+++ src/Detach.Tests/Tests/Collisions/Geometry3DTests.cs
@@ -1,5 +1,5 @@
 using Detach.Collisions;
-using Detach.Collisions.Primitives;
+using Detach.Collisions.Primitives3D;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Numerics;