+++ src/Detach/Collisions/Primitives2D/Circle.cs
@@ -1,6 +1,6 @@
 ﻿using System.Numerics;
 
-namespace Detach.Collisions.Primitives;
+namespace Detach.Collisions.Primitives2D;
 
 public record struct Circle
 {