+++ src/Detach/Collisions/Primitives2D/OrientedRectangle.cs
@@ -2,7 +2,7 @@
 using Detach.Numerics;
 using System.Numerics;
 
-namespace Detach.Collisions.Primitives;
+namespace Detach.Collisions.Primitives2D;
 
 public record struct OrientedRectangle
 {